requirements.txt
asgiref==3.5.0
attrs==21.4.0
cached-property==1.5.2
certifi==2021.10.8
charset-normalizer==2.0.11
Deprecated==1.2.13
Django==3.2.9
django-braces==1.15.0
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
isodate==0.6.1
kavenegar==1.1.2
lxml==4.7.1
packaging==21.3
platformdirs==2.5.0
psycopg2-binary==2.9.3
pyparsing==3.0.7
pytz==2021.3
requests==2.27.1
requests-file==1.5.1
requests-toolbelt==0.9.1
setuptools-scm==6.4.2
six==1.16.0
sqlparse==0.4.2
tomli==2.0.0
urllib3==1.26.8
wrapt==1.13.3
zeep==4.1.0
zope.event==4.5.0
zope.interface==5.4.0
